feat(azuredevops): OAuth, sync users and user deletion. (#466) (#470)
## Description

Azure DevOps Integration.
Context: 
- It provides multiple organizations. So we need to redirect a user to
select an organization he wants.
- To use an OAuth app, we need to set the organization-level policy
setting. So we need to redirect the user to the organization settings
page to set the option if it is OFF.

Implementation:
1. OAuth flow
2. Sync users
3. Delete a user

## Additional Notes

I implemented pagination feature based on the docs, but didn't test it
because the default is 100.

## Checklist:

Before you create this PR, confirm all the requirements listed below by
checking the checkboxes like this (`[x]`).

- [x] I have performed a self-review of my own code.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have added tests to cover the new feature or fixes.

## Description

Please include a summary of the changes and issues addressed in this PR.
Also, explain the motivation and context of each update. List any
dependencies or prerequisites needed for implementing this change.

## Additional Notes

Add any other context or screenshots about the feature request here.

## Checklist:

Before you create this PR, confirm all the requirements listed below by
checking the checkboxes like this (`[x]`).

- [ ] I have performed a self-review of my own code.
- [ ] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have added tests to cover the new feature or fixes.

---------

Co-authored-by: Maksym <148101061+mbliakharskyi@users.noreply.github.com>