feat(docusign): fetch organisations directly from the elba-api (#525)
## Description

This is a refactor of the docusign integration to fetch the organisation
from elba.
Previously docusign was using nango during the installation flow. Now
elba will handle this internally. As a consequence, the installation
frontend is not necessary, as well as the database as now it will
directly fetch the organisations from the elba api.

## Additional Notes

Add any other context or screenshots about the feature request here.

## Checklist:

Before you create this PR, confirm all the requirements listed below by
checking the checkboxes like this (`[x]`).

- [ ] I have performed a self-review of my own code.
- [ ] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have added tests to cover the new feature or fixes.