fix(slack): fix installation flow (#578)
## Description

This fixes Slack installation for the following cases:
- When an organisation was trying to install slack with a team that was
already in database for a different organisation, the different
organisation would still be tied to that team. Now the last organisation
will be tied to that team.
- When an organisation was trying to install slack with a team that is
different from the one it is tied to in database, the installation would
just fail. Now the previous team tied to the organisation is removed
from database before inserting the new one.

## Additional Notes

Add any other context or screenshots about the feature request here.

## Checklist:

Before you create this PR, confirm all the requirements listed below by
checking the checkboxes like this (`[x]`).

- [x] I have performed a self-review of my own code.
- [ ] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have added tests to cover the new feature or fixes.