chore(google): personal drive sync optimization (#621)
## Description

This improves Google data protection sync. We now use the `shared` file
attribute to know if a file is shared or not. This attribute only works
for personal drives. When the file is not shared, we avoid retrieving
its permissions and sending it. By consequence, personal drives sync
speed should be improved

## Additional Notes

Add any other context or screenshots about the feature request here.

## Checklist:

Before you create this PR, confirm all the requirements listed below by
checking the checkboxes like this (`[x]`).

- [x] I have performed a self-review of my own code.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have added tests to cover the new feature or fixes.